chore(components): remove the others folder
